feat: 제품 목록 dummy 데이터를 데이터 레이어에서 반환하도록 변경
