+++ README.md
@@ -17,7 +17,7 @@
 - [ ] 장바구니 담기 버튼 기능을 구현한다.
 - [x] 데이터 레이어 역할을 하는 패키지를 만들고, 힌트에서 주어진 Cart를 Room처럼 사용한다.
 - [x] Product에 id String을 추가한다.
-- [ ] dummy 데이터를 데이터 레이어에서 반환하도록 변경한다.
+- [x] dummy 데이터를 데이터 레이어에서 반환하도록 변경한다.
 - 장바구니 화면을 구현한다.
   - [ ] 담긴 상품의 수량을 조절할 수 있어야 한다.
   - [ ] 수량을 1보다 작게 하면 장바구니에서 상품이 제거된다.
